requirements.txt
openai
pymupdf
python-dotenv
paddleocr
pillow
numpy